+++ lib/flutter_html.dart
@@ -3,6 +3,7 @@ library flutter_html;
 import 'package:flutter/material.dart';
 import 'package:flutter_html/html_parser.dart';
 import 'package:flutter_html/style.dart';
+import 'package:webview_flutter/webview_flutter.dart';
 
 class Html extends StatelessWidget {
   /// The `Html` widget takes HTML as input and displays a RichText
@@ -40,6 +41,7 @@ class Html extends StatelessWidget {
     this.onImageTap,
     this.blacklistedElements = const [],
     this.style,
+    this.navigationDelegateForIframe,
   }) : super(key: key);
 
   final String data;
@@ -59,6 +61,11 @@ class Html extends StatelessWidget {
   /// Fancy New Parser parameters
   final Map<String, Style> style;
 
+  /// Decides how to handle a specific navigation request in the WebView of an
+  /// Iframe. It's necessary to use the webview_flutter package inside the app
+  /// to use NavigationDelegate.
+  final NavigationDelegate navigationDelegateForIframe;
+
   @override
   Widget build(BuildContext context) {
     final double width = shrinkWrap ? null : MediaQuery.of(context).size.width;
@@ -74,6 +81,7 @@ class Html extends StatelessWidget {
         style: style,
         customRender: customRender,
         blacklistedElements: blacklistedElements,
+        navigationDelegateForIframe: navigationDelegateForIframe,
       ),
     );
   }