+++ lib/src/html_elements.dart
@@ -30,8 +30,6 @@ const STYLED_ELEMENTS = [
   "strong",
   "sub",
   "sup",
-  "td",
-  "th",
   "time",
   "tt",
   "u",
@@ -68,6 +66,7 @@ const STYLED_ELEMENTS = [
   "p",
   "pre",
   "section",
+  "summary",
   "ul",
 ];
 
@@ -90,6 +89,7 @@ const REPLACED_ELEMENTS = [
 ];
 
 const LAYOUT_ELEMENTS = [
+  "details",
   "table",
   "tr",
   "tbody",
@@ -97,7 +97,9 @@ const LAYOUT_ELEMENTS = [
   "thead",
 ];
 
-const TABLE_STYLE_ELEMENTS = ["col", "colgroup"];
+const TABLE_CELL_ELEMENTS = ["th", "td"];
+
+const TABLE_DEFINITION_ELEMENTS = ["col", "colgroup"];
 
 /**
   Here is a list of elements with planned support: