+++ lib/src/layout_element.dart
@@ -1,8 +1,11 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_html/html_parser.dart';
 import 'package:flutter_html/src/html_elements.dart';
 import 'package:flutter_html/src/styled_element.dart';
 import 'package:flutter_html/style.dart';
+import 'package:flutter_layout_grid/flutter_layout_grid.dart';
 import 'package:html/dom.dart' as dom;
 
 /// A [LayoutElement] is an element that breaks the normal Inline flow of
@@ -28,51 +31,120 @@ class TableLayoutElement extends LayoutElement {
 
   @override
   Widget toWidget(RenderContext context) {
-    final colWidths = children
-        .where((c) => c.name == "colgroup")
-        .map((group) {
-          return group.children.where((c) => c.name == "col").map((c) {
-            final widthStr = c.attributes["width"] ?? "";
-            if (widthStr.endsWith("%")) {
-              final width =
-                  double.tryParse(widthStr.substring(0, widthStr.length - 1)) *
-                      0.01;
-              return FractionColumnWidth(width);
-            } else {
-              final width = double.tryParse(widthStr);
-              return width != null ? FixedColumnWidth(width) : null;
-            }
-          });
-        })
-        .expand((i) => i)
-        .toList()
-        .asMap();
-
-    return Container(
-        decoration: BoxDecoration(
-          color: style.backgroundColor,
-          border: style.border,
-        ),
-        width: style.width,
-        height: style.height,
-        child: Table(
-          columnWidths: colWidths,
-          children: children
-              .map((c) {
-                if (c is TableSectionLayoutElement) {
-                  return c.toTableRows(context);
+    final rows = <TableRowLayoutElement>[];
+    List<TrackSize> columnSizes = <TrackSize>[];
+    for (var child in children) {
+      if (child is TableStyleElement) {
+        // Map <col> tags to predetermined column track sizes
+        columnSizes = child.children
+            .where((c) => c.name == "col")
+            .map((c) {
+              final span =
+                  int.parse(c.attributes["span"] ?? "1", onError: (_) => 1);
+              final colWidth = c.attributes["width"];
+              return List.generate(span, (index) {
+                if (colWidth != null && colWidth.endsWith("%")) {
+                  final percentageSize = double.tryParse(
+                      colWidth.substring(0, colWidth.length - 1));
+                  return percentageSize != null
+                      ? FlexibleTrackSize(percentageSize * 0.01)
+                      : FlexibleTrackSize(1);
+                } else if (colWidth != null) {
+                  final fixedPxSize = double.tryParse(colWidth);
+                  return fixedPxSize != null
+                      ? FixedTrackSize(fixedPxSize)
+                      : FlexibleTrackSize(1);
+                } else {
+                  return FlexibleTrackSize(1);
                 }
-                return null;
-              })
-              .where((t) {
-                return t != null;
-              })
-              .toList()
-              .expand((i) => i)
-              .toList(),
-        ));
+              });
+            })
+            .expand((element) => element)
+            .toList(growable: false);
+      } else if (child is TableSectionLayoutElement) {
+        rows.addAll(child.children.whereType());
+      } else if (child is TableRowLayoutElement) {
+        rows.add(child);
+      }
+    }
+
+    // All table rows have a height intrinsic to their (spanned) contents
+    final rowSizes =
+        List.generate(rows.length, (_) => IntrinsicContentTrackSize());
+
+    // Calculate column bounds
+    int columnMax = rows
+        .map((row) => row.children
+            .whereType<TableCellElement>()
+            .fold(0, (int value, child) => value + child.colspan))
+        .fold(0, max);
+
+    // Place the cells in the rows/columns
+    final cells = <GridPlacement>[];
+    final columnRowOffset = List.generate(columnMax + 1, (_) => 0);
+    int rowi = 0;
+    for (var row in rows) {
+      int columni = 0;
+      for (var child in row.children) {
+        if (columnRowOffset[columni] > 0) {
+          columnRowOffset[columni] = columnRowOffset[columni] - 1;
+          columni++;
+        }
+        if (child is TableCellElement) {
+          cells.add(GridPlacement(
+            child: Container(
+              width: double.infinity,
+              padding: child.style.padding ?? row.style.padding,
+              decoration: BoxDecoration(
+                color: child.style.backgroundColor ?? row.style.backgroundColor,
+                border: child.style.border ?? row.style.border,
+              ),
+              child: SizedBox.expand(
+                child: Container(
+                  alignment: child.style.alignment ?? style.alignment ??
+                      Alignment.centerLeft,
+                  child: StyledText(
+                    textSpan: context.parser.parseTree(context, child),
+                    style: child.style,
+                  ),
+                ),
+              ),
+            ),
+            columnStart: columni,
+            columnSpan: child.colspan,
+            rowStart: rowi,
+            rowSpan: child.rowspan,
+          ));
+          columnRowOffset[columni] = child.rowspan - 1;
+          columni += child.colspan;
+        }
+      }
+      rowi++;
+    }
+
+    // Create column tracks (insofar there were no colgroups that already defined them)
+    List<TrackSize> finalColumnSizes = (columnSizes ?? <TrackSize>[]).take(
+        columnMax).toList();
+    finalColumnSizes += List.generate(
+        max(0, columnMax - finalColumnSizes.length),
+            (_) => FlexibleTrackSize(1));
+    return Container(
+      decoration: BoxDecoration(
+        color: style.backgroundColor,
+        border: style.border,
+      ),
+      width: style.width,
+      height: style.height,
+      child: LayoutGrid(
+        gridFit: GridFit.loose,
+        templateColumnSizes: finalColumnSizes,
+        templateRowSizes: rowSizes,
+        children: cells,
+      ),
+    );
   }
 }
+
 
 class TableSectionLayoutElement extends LayoutElement {
   TableSectionLayoutElement({
@@ -82,18 +154,8 @@ class TableSectionLayoutElement extends LayoutElement {
 
   @override
   Widget toWidget(RenderContext context) {
+    // Not rendered; TableLayoutElement will instead consume its children
     return Container(child: Text("TABLE SECTION"));
-  }
-
-  List<TableRow> toTableRows(RenderContext context) {
-    return children.map((c) {
-      if (c is TableRowLayoutElement) {
-        return c.toTableRow(context);
-      }
-      return null;
-    }).where((t) {
-      return t != null;
-    }).toList();
   }
 }
 
@@ -106,35 +168,55 @@ class TableRowLayoutElement extends LayoutElement {
 
   @override
   Widget toWidget(RenderContext context) {
+    // Not rendered; TableLayoutElement will instead consume its children
     return Container(child: Text("TABLE ROW"));
   }
+}
 
-  TableRow toTableRow(RenderContext context) {
-    return TableRow(
-        decoration: BoxDecoration(
-          border: style.border,
-          color: style.backgroundColor,
-        ),
-        children: children
-            .map((c) {
-              if (c is StyledElement && c.name == 'td' || c.name == 'th') {
-                return TableCell(
-                    child: Container(
-                        padding: c.style.padding,
-                        decoration: BoxDecoration(
-                          color: c.style.backgroundColor,
-                          border: c.style.border,
-                        ),
-                        child: StyledText(
-                          textSpan: context.parser.parseTree(context, c),
-                          style: c.style,
-                        )));
-              }
-              return null;
-            })
-            .where((c) => c != null)
-            .toList());
+class TableCellElement extends StyledElement {
+  int colspan = 1;
+  int rowspan = 1;
+
+  TableCellElement({
+    String name,
+    String elementId,
+    List<String> elementClasses,
+    @required List<StyledElement> children,
+    Style style,
+    dom.Element node,
+  }) : super(
+      name: name,
+      elementId: elementId,
+      elementClasses: elementClasses,
+      children: children,
+      style: style,
+      node: node) {
+    colspan = _parseSpan(this, "colspan");
+    rowspan = _parseSpan(this, "rowspan");
   }
+
+  static int _parseSpan(StyledElement element, String attributeName) {
+    final spanValue = element.attributes[attributeName];
+    return spanValue == null ? 1 : int.tryParse(spanValue) ?? 1;
+  }
+}
+
+TableCellElement parseTableCellElement(dom.Element element,
+    List<StyledElement> children,
+) {
+  final cell = TableCellElement(
+    name: element.localName,
+    elementId: element.id,
+    elementClasses: element.classes.toList(),
+    children: children,
+    node: element,
+  );
+  if (element.localName == "th") {
+    cell.style = Style(
+      fontWeight: FontWeight.bold,
+    );
+  }
+  return cell;
 }
 
 class TableStyleElement extends StyledElement {
@@ -146,9 +228,8 @@ class TableStyleElement extends StyledElement {
   }) : super(name: name, children: children, style: style, node: node);
 }
 
-TableStyleElement parseTableDefinitionElement(
-  dom.Element element,
-  List<StyledElement> children,
+TableStyleElement parseTableDefinitionElement(dom.Element element,
+    List<StyledElement> children,
 ) {
   switch (element.localName) {
     case "colgroup":
@@ -163,11 +244,74 @@ TableStyleElement parseTableDefinitionElement(
   }
 }
 
+class DetailsContentElement extends LayoutElement {
+  List<dom.Element> elementList;
+
+  DetailsContentElement({
+    String name,
+    List<StyledElement> children,
+    dom.Element node,
+    this.elementList,
+  }) : super(name: name, node: node, children: children);
+
+  @override
+  Widget toWidget(RenderContext context) {
+    return ExpansionTile(
+        title: children != null && elementList.first.localName == "summary" ? StyledText(
+          textSpan: TextSpan(
+            style: style.generateTextStyle(),
+            children: [children
+                .map((tree) => context.parser.parseTree(context, tree))
+                .toList().first] ??
+                [],
+          ),
+          style: style,
+        ) : Text("Details"),
+        children: [
+          StyledText(
+            textSpan: TextSpan(
+              style: style.generateTextStyle(),
+              children: getChildren(children, context)
+            ),
+            style: style,
+          ),
+        ]
+    );
+  }
+
+  List<InlineSpan> getChildren(List<StyledElement> children, RenderContext context) {
+    if (children.map((tree) => context.parser.parseTree(context, tree)).toList() == null) {
+      return [];
+    } else {
+      List<InlineSpan> reducedChildren = children.map((tree) => context.parser.parseTree(context, tree)).toList();
+      reducedChildren.removeAt(0);
+      return reducedChildren;
+    }
+  }
+}
+
+class EmptyLayoutElement extends LayoutElement {
+  EmptyLayoutElement({String name = "empty"}) : super(name: name);
+
+  @override
+  Widget toWidget(_) => null;
+}
+
 LayoutElement parseLayoutElement(
-  dom.Element element,
-  List<StyledElement> children,
+    dom.Element element,
+    List<StyledElement> children,
 ) {
   switch (element.localName) {
+    case "details":
+      if (children == null || children.isEmpty) {
+        return EmptyLayoutElement();
+      }
+      return DetailsContentElement(
+          node: element,
+          name: element.localName,
+          children: children,
+          elementList: element.children
+      );
     case "table":
       return TableLayoutElement(
         name: element.localName,