+++ lib/src/styled_element.dart
@@ -25,9 +25,11 @@ class StyledElement {
   bool matchesSelector(String selector) =>
       _node != null && matches(_node, selector);
 
-  Map<String, String> get attributes => _node.attributes.map((key, value) {
+  Map<String, String> get attributes =>
+      _node?.attributes?.map((key, value) {
         return MapEntry(key, value);
-      });
+      }) ??
+      Map<String, String>();
 
   dom.Element get element => _node;
 
@@ -349,8 +351,6 @@ StyledElement parseStyledElement(
         verticalAlign: VerticalAlign.SUPER,
       );
       break;
-    case "th":
-      continue bold;
     case "tt":
       continue monospace;
     underline: